backend/app/lit_protocol.py
"""
Lit Protocol integration for encryption/decryption
"""

import os
from typing import Dict, Any
from lit_sdk import LitProtocol


class LitEncryption:
    """Wrapper for Lit Protocol encryption operations"""

    def __init__(self):
        self.lit = None
        self._initialized = False

    async def _initialize(self):
        """Initialize Lit Protocol connection"""
        if self._initialized:
            return

        try:
            # Initialize Lit Protocol SDK
            # Note: Actual implementation would use proper Lit SDK initialization
            self.lit = LitProtocol()
            await self.lit.connect()
            self._initialized = True
        except Exception as e:
            print(f"Warning: Lit Protocol initialization failed: {e}")
            # In development, we might want to continue without Lit
            self._initialized = False

    async def encrypt(
        self, data: Dict[str, Any], access_conditions: Dict[str, Any]
    ) -> str:
        """
        Encrypt data using Lit Protocol

        Args:
            data: Data to encrypt
            access_conditions: Access control conditions for decryption

        Returns:
            Encrypted data as string
        """
        await self._initialize()

        if not self._initialized:
            # Fallback for development
            import json
            import base64

            data_str = json.dumps(data)
            return base64.b64encode(data_str.encode()).decode()

        try:
            # Actual Lit Protocol encryption
            # This is a placeholder - actual implementation would use Lit SDK
            encrypted = await self.lit.encrypt(
                data=data, access_conditions=access_conditions
            )
            return encrypted
        except Exception as e:
            raise Exception(f"Encryption failed: {e}")

    async def decrypt(self, encrypted_data: str, access_conditions: Dict[str, Any]) -> Dict[str, Any]:
        """
        Decrypt data using Lit Protocol

        Args:
            encrypted_data: Encrypted data string
            access_conditions: Access control conditions

        Returns:
            Decrypted data dictionary
        """
        await self._initialize()

        if not self._initialized:
            # Fallback for development
            import json
            import base64

            data_str = base64.b64decode(encrypted_data.encode()).decode()
            return json.loads(data_str)

        try:
            # Actual Lit Protocol decryption
            decrypted = await self.lit.decrypt(
                encrypted_data=encrypted_data, access_conditions=access_conditions
            )
            return decrypted
        except Exception as e:
            raise Exception(f"Decryption failed: {e}")

    def is_connected(self) -> bool:
        """Check if Lit Protocol is connected"""
        return self._initialized and self.lit is not None

