backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
browser-use==0.1.0
playwright==1.40.0
langchain-openai==0.0.5
python-multipart==0.0.6
web3==6.11.0
pyotp==2.9.0
qrcode==7.4.2
resend==0.7.0
sendgrid==6.11.0
websockets==12.0
pillow==10.1.0
